+++ .env.example
@@ -1,1 +1,1 @@
-PORT=4000
+WS_PORT=4000
