+++ package.json
@@ -2,11 +2,11 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
+        "start": "node ./src/index.ts",
+        "start:dev": "tsx watch ./src/index.ts",
+        "format": "eslint ./src --fix",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "author": "Andrei Auchynnikau",
@@ -14,6 +14,7 @@
     "license": "ISC",
     "devDependencies": {
         "@types/node": "^20.11.18",
+        "@types/ws": "^8.5.10",
         "@typescript-eslint/eslint-plugin": "^6.21.0",
         "@typescript-eslint/parser": "^6.21.0",
         "cross-env": "^7.0.3",
@@ -22,9 +23,9 @@
         "eslint-config-airbnb-typescript": "^17.1.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-prettier": "^5.1.3",
-        "nodemon": "^2.0.16",
         "prettier": "^3.2.5",
         "ts-node-dev": "^2.0.0",
+        "tsx": "^4.7.1",
         "typescript": "^5.3.3"
     },
     "dependencies": {
