+++ src/ws_server/index.ts
@@ -1,0 +1,3 @@
+export const handleWsConnection = () => {
+
+};
